+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0]
+
+Required breaking changes to support Ferrite v1 and later.
+### Changed
+ - Keyword arg of `get_ferrite_grid`: `generate_facesets` -> `generate_facetsets`
+ - Function renamed: `create_faceset` -> `create_facetset`
 
 ## [0.1.8]